--- python/bloqade/lanes/dialects/__init__.py
@@ -1,1 +1,0 @@
-from bloqade.lanes.dialects import place as place
